Rename ListPage to ServerPage.
